requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2 
sec_api
marker-pdf
praw
langchain
pdf-kit
# finnlp
akshare
tushare
finnhub-python
parsel
requests
pandas
tqdm
pytz
unstructured==0.5.6
faker
ratelimit
yfinance
